refactor: 프로젝트 리뷰 api 폴더 정리
